Moved remaining function to separate file, and added typing
